:bento: Add accent color and font definitions
The accent color has now been defined from the official style guide
of the project. Likewise, the Xcode definitions for the Salmon font
is in place, though the font files themselves are not committed to
be consistent with the other game source.

Finally, the Bunker and CranberrySprite packages have been added as
dependencies to the project.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>