:lipstick: Add font family extensions
These extension methods are used to pick the appropriate font with
the Salmon typeface, based on the specifications in the game's
style guide.

Currently, the salmonMonospaced function doesn't work completely
as intended, but this will hopefully be changed in the future once
a means of determining a font's size from the Font struct is
determined.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>