:lipstick: Create main menu, update window
A prototype main menu is in place, which displays the game's name
and provides an option to start the demo scene. This menu now works
across iOS and macOS with some aspect ratio constraints (16/9).

Additionally, the Entrypoint now contains app delegates on the iOS
and macOS side for specific operations that the SwiftUI app lifecycle
may not be ready for, such as authenticating with Game Center and
forcing the aspect ratio of a window when resizing.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>