:sparkles: Add Game Center authentication
Game Center is now registered with the game, and players can sign
in to Game Center through the game. Additionally, the access point
is now visible.

To do this, the AppDelegate needed some slight modifications to
account for the different view controller types across platforms
when the authentication handler closure executes. AppDelegate was
refactored to utilize extensions to share common methods, and it
has been pulled into its own file instead of sitting in Entrypoint.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>