:heavy_plus_sign: Add JensonKit
JensonKit is the internal framework used to handle dialogue. This
is the equivalent to Dialogic for the Swift/Xcode realm, though
there doesn't exist a graphical editor at this moment.

Marteau already handles conversions to the Jenson file format, and
this framework includes a reader to read and decode the format.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>