:construction: Configure paintbrush scene
The PaintbrushScene should now be able to properly listen for panel
drawing events and validating with the prototype model. The code is
similarly structured to that of PaintbrushTest with some minor
tweaks.

Future commits should likely focus on the actual design of the scene
and how to implement loading in paintings.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>