:construction: Load configs and highlighing
New delegate protocols have been added to handle reading from a
configuration file in JSON, which contains information about the
puzzles in a given stage, as well as their palettes. This should
help with getting puzzles configured properly and validating with
the correct puzzle set.

This code hasn't been completely tested yet, but this should be
worked on in future commits while the focus shifts to the level
scenes themselves.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>