:construction: Create and finalize first scene
A new demo scene for Paintbrush exists on the main menu as the
"Run Paintbrush Demo" option. It loads in the prototype stage with
the first puzzle.

It appears that the validation scheme doesn't run in the Simulator.
However, it works correctly on device. To make the prototype work
in the simulator, the actual result is ignored in the highlight
process and alway returns true. This behavior is automatically
disabled on device with a compiler flag.

The next step will be finding a means to move between puzzles in
the prototype to demonstrate the modularity.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>