:children_crossing: Overhaul Paintbrush validation
Further testing with the Paintbrush system has shown that the new
system, updated in commit 2167c3e, was still flawed in nature. Most
notably, some lines were not being registered correctly and would
never reach a viable solution.

This commit attempts to address these issues with several major
changes.

First, the model has been retrained with new hyperparameters and
data. The dataset now contains 900 training images and 54 testing
images, recorded from previous sessions in-game. Furthermore, the
following settings have been applied:

- Add noise
- Expose
- Removed crop

This resulted in a well-improved model, with 100% accuracy in
training and validation. Of course, it is likely that the model has
overfitted to some degree, as it occassionally misses the mark. But,
this is a much better model than what was previously there.

Next, the threshold limit has been removed from the PaintbrushSolver
protocol, as some guesses would still appear with lower scores and
get cut off. Likewise, the scene no longer handles this.

Finally, to "progress" through levels, the debugging sprite used to
test the screenshot mechanisms in-game can now be tapped on to move
to the next level, regardless if it was solved before. This will be
removed in a later commit when a better system is in place to keep
track of states in-game.

On another note, this commit also introduces a new Info.plist key
called 'PaintbrushRecourdOutput'. This key controls whether the
player's drawings get saved to disk for debugging purposes. This is
controlled with the build setting PNT_RECORD_OUTPUT and is on by
default for debugging builds.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>