:boom: Restructure codebase
The codebase has been completely restructured to be easier to
navigate, and it is based around a module system.

The definition logic for the Paintbrush puzzle system has been moved
to the Paintbrush Swift package, which is then added to the game.

The game's source code has been moved into the Shounin group and
added as a project to the new Indexing Your Heart workspace.

Hopefully, this new structure will keep things organized and
working as intended.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>