:construction: Connect Paintbrush to Shounin
The testing world is now able to launch into a puzzle and save the
state of the world from a previous run. The player can solve the
puzzle configured and exit out by tapping the button on the bottom
right hand corner.

If a puzzle remains unsolved or was incorrect, the player can return
to that puzzle to solve it. Solved puzzles can't be interacted with
again, as there is no system to "remember" what the player drew yet.

Furthermore, the system isn't completely ready for a save system or
multiple puzzles. More work will need to be done to ensure this case.

The next commits will focus on handling the metapuzzle stage and the
transition to the corresponding Caslon scene for that chapter.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>