:construction: Test multipuzzle, solved states
I've confirmed that the code I have written for handling multiple
puzzles works as intended. Of course, this doesn't account for the
metapuzzle, but I will work on this for the next commit.

Likewise, the Paintbrush scene has been tweaked to allow revisiting
when the panel has been solved. However, to restore the solved state
from before, an image located in the Documents directory of the
sandbox is loaded into the scene if one is present. Furthermore,
when the puzzle is solved and the player leaves the scene, the panel
is saved to an image in the same place, reutilizing the code for
creating debugging images with the ML model.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>