:children_crossing: Finishing touches
In this commit, I've made some finishing touches to help advance
the development after the prototype launch.

First, I've created a system to load in a specific set of scenes
in a given order with a JSON configuration that gets loaded at
launch time. This will control which levels play in a specific
order.

Next, I've tested the prototype on iOS and fixed any remaining
snags that may make the game unplayable on iPad and iPhone.

Finally, I've fixed the story conversion script in the build config
so that it executes to the new structure.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>