:sparkles: Call next after Caslon scene ends
This change effectively allows the game to move to the next stage,
or quit the game if there is nothing left. It is unknown if this
change actually moves to the next scene and reloads the view, but
it will be interesting to see how it develops.

For the case of HOH-7 being fulfilled, this commit resolves HOH-7.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>