:memo: Update README
The intro paragraph in the README now more closely follows the
description found on the game's website. Additionally, the bug report
section has been altered to let others know about the Discord server
or send via email. This change occurred because the tracking system
has been migrated to Jira in a private instance.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>