:recycle: Refactor Paintbrush scene for Protractor
Paintbrush now conforms to the PaintbrushProtractorSolver protocol
and acts as a solver using the Protractor gesture algorithm. Lines
were drawn in-game and kept as their SpriteKit coordinates to
maintain consistency across a wide variety of device sizes while
still having valid coordinates for recognition.

Before and after recognition, the points are translated to the
origin of the scene to be consistent, so recognition can still
apply, even if the panel is not in the center of the scene.

All CoreML-specific code has been removed, as the game will no
longer support using machine learning to validate lines.

This should effectively resolve HOH-4.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>