:children_crossing: Show tutorial in Paintbrush
On puzzles that define the "show_tutorial_hint" key in their config,
a tutorial node will be present in the scene. This node shows a
finger or cursor and shows the player to swipe across the canvas
to solve the first puzzle in the game.

This is the first of multiple commits to create a proper interactive
tutorial for playing the game. The next commits will focus on the
movement mechanics.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>