:children_crossing: HOH-6 Show movement tutorial
The game environment scene will now show a tutorial for movement
when a stage is defined to have the "show_tutorials" key set to
true. For macOS, this shows the keyboard controls. On iOS and
iPadOS, this shows the touch controls.

The next commit will focus on showing how to invoke solve mode.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>