:children_crossing: Add Caslon tutorial hints
Caslon scenes now include a tutorial hint that shows how to move
to the next step in the story. On Mac, this displays a cursor; for
iPhone and iPad, this displays a finger tap icon. These icons are
also used in the game environment's tutorial.

As per the tutorial mockups, this should complete the tutorial
issue and thus resolve HOH-6.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>