:bento: Add character sprites to prototype
This commit adds some prototype character cutouts for the first
chapter of the game. This allows for the game to display the
characters on screen, triggered by refresh events.

The Caslon scene has been modified a little to include more layers.
New layers have been added to put images on the left and right sides.
Values ending in '01' correspond to the left, and values ending in
'02' correspond to the right.

Finally, the chapter script for Chapter One has been updated to
add the refresh events for showing the sprites on the screen.

Hopefully, in the future, the prototype sprites can be replaced
with artwork created by Kabora or other artists.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>