:children_crossing: Add dust particles
The game environment now has an emitter that will create dust
particles that follow the player's camera to add more variety to
the environment.

Additionally, the metapuzzle template (hoh_1i) has been redone to
better closely match the intended solution, and the resampling
count in the recognition algorithm has increased to 36.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>