:sparkles: Add pathfinding to player navigation
For the iOS version, the player will now move along a path using
a pathfinding algorithm to connect from one part of the map to
another. This utilizes a combination of SKTiled and GameplayKit to
achieve this behavior.

This allows for the player's animation sprites to play during the
movement sequences.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>