:recycle: Clean up codebase
The codebase has been cleaned up to make reading through the game
environment easier. The game environment uses a new environment
delegate to handle what the environment was originally responsible
for, such as loading puzzles and displaying the tutorial.

Although GameEnvironment itself currently conforms to the new
GameEnvironmentDelegate protocol, this can be refactored to have
the logic separated, when necessary. Theoretically, this should also
allow for unit tests to be written by adding a fake delegate.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>