:sparkles: Add footsteps
Finally, we have footstep sounds in the game that trigger when the
player moves around in the world.

I've also removed the ambient sounds for the panels in the main
environment, as it did result in significant memory increase. I will
investigate how to go about this better in the next sprint, M6.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>