:bug: Use SKAudioNodes
There appears to be a memory leak issue when using the default
SKAction.playSoundFileNamed, never freeing the audio player when
finished.

To work around this, code using this call has been changed to a
more proper audio node system thanks to the convenience initializers
and SKAudioNode extensions the project contains. Memory usage now
sits at around 90MB in the environment, fluctuating when moving from
scene to scene.

This should effectively solve HOH-11, and it may also carry the
bulk for memory optimizations.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>