:bug: Prevent a crash caused in Paintbrush
The Protractor algorithm used in Paintbrush requires a minimum of 16 pts
to resample and vectorize a gesture. However, a player may try to draw a
dot, either by accident or with intent.

A check has been added to `resampled` to ensure that there exists the
minimum number of points to perform a successful resampling.