:recycle: Address memory leaks, load times
In a recent iOS update, the scene that was being cached in the observed
app state was not being properly set up, warranting a new approach
where the scene would be rebuilt from a small structure containing info
about the state of the environment such as the player's position, the
stage name, and what puzzles were solved already. While this does cause
a temporary spike in memory usage, the spike dissipates once the
originating Paintbrush scene is deallocated from memory, so it's not as
noticeable.

Additionally, the game will try to asynchronously load in the tilesets
for the maps in the background while the game first starts up. This
allows for faster load times of the game environment when being
regenerated, since it doesn't need to read the tilesets again to render
the map.