:children_crossing: General QoL
Disables the player's animation tree when in the editor (so as to not
require saving the world on open... every time).

Changes the signal registration from the numpad field to not
over-register.

Pressing the "cancel" key (ESC) will close the numpad field.

Changes the player's Y-sort origin to work better with the world's tile
maps (thanks, Maxi).

Updates the Y-sort origins for various tiles, creating for a more
cohesive experience (thanks, Maxi).

Includes content for the environmental puzzle for the numbers section.

Disables touch emulation on the main project's settings.