:construction: Create an exportable iOS project
This commit sets up an iOS project, ready for exporting easily to a
physical device or to the App Store. Instructions are included to better
describe how to use it to create a build.

This commit also updates SwiftGodotCore to address the issues for the
framework on iOS.

Finally, this commit removes the old dep-based recipes in the Justfile,
preferring extensions.