:sparkles: Save the game when entering background
Godot doesn't come with a built-in solution to listen for app lifecycle
events, such as when the app enters the background. However, we can
leverage the iOS SDK and listen for these events manually with the
NotificationCenter. Whenever such events are emitted from
NotificationCenter, the Rollinsport message bus will send
Godot-compatible signals that can be used to handle events in Godot.

This adapter is the driver behind this feature, though it only works on
iOS or places where UIKit can be imported.