Create C# solution
Besides the main game (Shounin) having a new C# project, a new
Jenson.NET package was created, along with unit tests, to start work on
Jenson v3 with KDL. JensonKit will need to be updated in the future to
support this new format.

This commit also removes some Nova tasks that can't be cloned on
Windows. These will likely need to be recreated for .NET, anyway. The
.gitignore was updated to also include files ignored for C# and .NET
projects.