IOH-50: Recreate word puzzles
Similar to the number puzzle nodes, everything functions except for
saving and prefilling puzzle data after the puzzle has been solved.

An important discovery that needs to be addressed soon is that the
visual for the cylinder is spelled incorrectly. It should be spelled
'kabutui', but the pixel art shows 'kabatui'.