IOH-50: Interchange with word puzzles
As with the previous commit, word puzzles now exchange messages with the
RollinsportMessageBus to save and load puzzle solutions. This comes with
the caveat that the prefill method for the puzzle text field might be
inaccurate, since the expected solution value is being mapped to both
the underlying field and the visible text label, which might not have
the correct formatting. This was also present in the previous iterations
of the game.