Store environment info in Playerfile
The environment contains game version information and the platform to
make debugging easier. When players are encountering issues with the
game, they can upload a copy of their Playerfile to help diagnose the
issue.

It is important to note that the game does not use the environment info.
The data is write-only, and it should be kept this way to not track
players.