:sparkles: Play preamble before starting the game
The message bus and Playerfile have been updated to keep track of which
scripts have been loaded, and the timeline can now be started after load
with a StartTimeline method.

Animations are a little jank right now, but I'm planning to create some
UI animation system not that different from UIKit to let this work
as I intend.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>