:sparkles: All choice questions, skip animations
In the dialogue node, you can provide an 'animated' property to specify
whether an animation should play. By default, this is assumed to be true
in Jenson.NET and the timeline. Setting this to false bypasses the
animations.

Questions can optionally provide a 'forceAll' property to denote whether
all choices need to be selected before finishing the question and moving
further in the timeline. Internally, the timeline will re-insert a
question for reuse, moving along once the choice lookup table is
emptied.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>